Change QMessageBox::Discard button to QMessageBox::Close in the start screen UI
Under many platforms the button is displayed as 'Discard', which is acceptable, but under GNOME
it is displayed as 'Close without Saving', which did not make any sense. The Qt documentation
mentions that this button could also be 'Don't Save' on some other platforms, which would also
not make sense.