Add support for power management plugins, and separate out KDE pm code to a plugin
Separating the KDE pm code as a plugin allows o-p-p to be always packaged with the plugin, and
defers the choice of using the code to runtime, avoiding the need for building and distributing
two versions, with and without KDE support.

The KDE pm plugin is built completely separately, and can be added to a o-p-p installation at
any time. (Currently only qmake build support is provided.)